ERR: erroneous cast to double caused inconsistency with sister image filters.

+++ Code/BasicFilters/itkAcosImageFilter.h
@@ -37,7 +37,7 @@ namespace function {
     ~Acos() {};
     inline TOutput operator()( const TInput & A )
     {
-      return (TOutput)acos((double) A);
+      return (TOutput)acos(A);
     }
   };
